declare function as float, not double.

+++ usr.bin/f77/libF77/r_abs.c
@@ -3,10 +3,10 @@
  * All rights reserved.  The Berkeley software License Agreement
  * specifies the terms and conditions for redistribution.
  *
- *	@(#)r_abs.c	5.1	6/7/85
+ *	@(#)r_abs.c	5.2	7/8/85
  */
 
-double r_abs(x)
+float r_abs(x)
 float *x;
 {
 if(*x >= 0)
